fix(payments): support decimal order quantities

+++ Core/Models/OrderItem.cs
@@ -23,7 +23,7 @@ public class OrderItem
     /// <summary>
     /// Discount in percentages for <see cref="OrderItem"/>
     /// </summary>
-    public int Discount { get; set; }
+    public decimal Discount { get; set; }
 
     /// <summary>
     /// 
@@ -33,5 +33,5 @@ public class OrderItem
     /// <summary>
     /// 
     /// </summary>
-    public int Quantity { get; set; }
+    public decimal Quantity { get; set; }
 }
